feat(ui): split CustomizePageUtils.ts into 3 focused modules (#28937)
* feat(ui): split CustomizePageUtils.ts into 3 focused modules

Closes open-metadata/openmetadata-collate#4230

Splits 830-line CustomizePageUtils.ts (with 25 eager ClassBase imports)
into three focused files to eliminate unnecessary bundle weight on every
entity detail page:

- CustomizePageEntityTabUtils.ts  — tab display helpers (no ClassBase deps)
  sortTabs, getTabDisplayName, getTabLabelMapFromTabs,
  checkIfExpandViewSupported, getDetailsTabWithNewLabel

- CustomizePageDispatchUtils.ts   — dispatch/admin flow helpers (keeps all
  25 ClassBase singletons; only imported by admin/customize pages)
  getDefaultTabs, getDefaultWidgetForTab, getCustomizableWidgetByPage,
  getDummyDataByPage, getWidgetsFromKey, getWidgetHeight,
  getGlossaryTermDefaultTabs, getGlossaryDefaultTabs

- CustomizePageWidgetUtils.ts     — widget/layout helpers (no ClassBase deps)
  getAddWidgetHandler, updateWidgetHeightRecursively,
  getLayoutFromCustomizedPage, asyncNoop

CustomizePageUtils.ts deleted; all ~50 consumer files updated to import
directly from the correct split module. No re-export facade.

Test mocks updated: removed stale mock blocks on wrong paths (previously
mocking CustomizePageUtils which had no effect once consumers moved to
CustomizePagePureUtils); added correct mocks on CustomizePagePureUtils
for FileClassBase and DirectoryClassBase tests.


* fix checkstyle

---------
