+++ openmetadata-ui/src/main/resources/ui/src/components/Customization/CustomizeTabWidget/CustomizeTabWidget.tsx
@@ -37,12 +37,12 @@ import {
   getUniqueFilteredLayout,
 } from '../../../utils/CustomizableLandingPageUtils';
 import {
-  getAddWidgetHandler,
   getCustomizableWidgetByPage,
   getDefaultTabs,
   getDefaultWidgetForTab,
-  getTabDisplayName,
-} from '../../../utils/CustomizePage/CustomizePageUtils';
+} from '../../../utils/CustomizePage/CustomizePageDispatchUtils';
+import { getTabDisplayName } from '../../../utils/CustomizePage/CustomizePageEntityTabUtils';
+import { getAddWidgetHandler } from '../../../utils/CustomizePage/CustomizePageWidgetUtils';
 import { getEntityName } from '../../../utils/EntityNameUtils';
 import { TabItem } from '../../common/DraggableTabs/DraggableTabs';
 import AddDetailsPageWidgetModal from '../../MyData/CustomizableComponents/AddDetailsPageWidgetModal/AddDetailsPageWidgetModal';
