+++ openmetadata-ui/src/main/resources/ui/src/components/Customization/GenericTab/GenericTab.tsx
@@ -17,7 +17,7 @@ import { DetailPageWidgetKeys } from '../../../enums/CustomizeDetailPage.enum';
 import { PageType } from '../../../generated/system/ui/page';
 import { useGridLayoutDirection } from '../../../hooks/useGridLayoutDirection';
 import { WidgetConfig } from '../../../pages/CustomizablePage/CustomizablePage.interface';
-import { getWidgetsFromKey } from '../../../utils/CustomizePage/CustomizePageUtils';
+import { getWidgetsFromKey } from '../../../utils/CustomizePage/CustomizePageDispatchUtils';
 import { useGenericContext } from '../GenericProvider/GenericProvider';
 import { DynamicHeightWidget } from './DynamicHeightWidget';
 import './generic-tab.less';
