+++ openmetadata-ui/src/main/resources/ui/src/components/Customization/GenericWidget/GenericWidget.tsx
@@ -20,7 +20,7 @@ import { PageType } from '../../../generated/system/ui/page';
 import { WidgetCommonProps } from '../../../pages/CustomizablePage/CustomizablePage.interface';
 import { useCustomizeStore } from '../../../pages/CustomizablePage/CustomizeStore';
 import customizeGlossaryTermPageClassBase from '../../../utils/CustomizeGlossaryTerm/CustomizeGlossaryTermBaseClass';
-import { getDummyDataByPage } from '../../../utils/CustomizePage/CustomizePageUtils';
+import { getDummyDataByPage } from '../../../utils/CustomizePage/CustomizePageDispatchUtils';
 import { WIDGET_COMPONENTS } from '../../../utils/GenericWidget/GenericWidgetUtils';
 import { DEFAULT_ENTITY_PERMISSION } from '../../../utils/PermissionsUtils';
 import { EntityUnion } from '../../Explore/ExplorePage.interface';
