+++ openmetadata-ui/src/main/resources/ui/src/components/Dashboard/DataModel/DataModels/DataModelDetails.component.test.tsx
@@ -112,11 +112,14 @@ jest.mock('../../../../utils/DashboardDataModelClassBase', () => ({
   },
 }));
 
-jest.mock('../../../../utils/CustomizePage/CustomizePageUtils', () => ({
-  getTabLabelMapFromTabs: jest.fn().mockReturnValue({}),
-  getDetailsTabWithNewLabel: jest.fn().mockReturnValue([]),
-  checkIfExpandViewSupported: jest.fn().mockReturnValue(false),
-}));
+jest.mock(
+  '../../../../utils/CustomizePage/CustomizePageEntityTabUtils',
+  () => ({
+    getTabLabelMapFromTabs: jest.fn().mockReturnValue({}),
+    getDetailsTabWithNewLabel: jest.fn().mockReturnValue([]),
+    checkIfExpandViewSupported: jest.fn().mockReturnValue(false),
+  })
+);
 
 describe('DataModelDetails component', () => {
   it('should render successfully', () => {
