+++ openmetadata-ui/src/main/resources/ui/src/pages/CustomizableDataProductPage/CustomizableDataProductPage.tsx
@@ -27,7 +27,7 @@ import { DataProduct } from '../../generated/entity/domains/dataProduct';
 import { Page } from '../../generated/system/ui/page';
 import { PageType } from '../../generated/system/ui/uiCustomization';
 import { useMarketplaceStore } from '../../hooks/useMarketplaceStore';
-import { getDummyDataByPage } from '../../utils/CustomizePage/CustomizePageUtils';
+import { getDummyDataByPage } from '../../utils/CustomizePage/CustomizePageDispatchUtils';
 import { getEntityName } from '../../utils/EntityNameUtils';
 import { useCustomizeStore } from '../CustomizablePage/CustomizeStore';
 import '../CustomizeDetailsPage/customize-details-page.less';
