+++ openmetadata-ui/src/main/resources/ui/src/utils/Glossary/GlossaryTermClassBase.test.ts
@@ -25,10 +25,6 @@ jest.mock('./GlossaryTermUtils', () => ({
     .mockReturnValue([{ key: 'mock-tab' }]),
 }));
 
-jest.mock('../../utils/CustomizePage/CustomizePageUtils', () => ({
-  getTabLabelFromId: jest.fn((tab: string) => tab),
-}));
-
 jest.mock(
   '../../components/DataQuality/DataQualityDashboard/DataQualityDashboard.component',
   () => ({ __esModule: true, default: () => null })
